test(adapter): extract projection test module
